nc-passwd: clarify usage
Found a few reference https://github.com/nextcloud/nextcloudpi/issues/155
to issues with special characters in passwords, also with german umlaut ".
Which ones are sanitized? and need to be added to "info" or should it be in "description"?
I will also add it to Config-Reference, once established which characters to avoid ;-)