Don't fail ncp-backup-auto
Make sure the script don't return an exit status different from 0 so the
cron job doesn't send e-mails on failure, as notifications are being
delivered as Nextcloud notifications. Also redirect all output to a log
file, so in case of failure it is possible to review what happened.

This also makes the notification a bit more informative, as it says
which script failed.

The log files is overwritten on each run. This should not be a problem
since backups are run more frequently than once per day, so there should
be a fair amount of time to see the backup logs.

Signed-off-by: Leandro Lucarella <luca@llucax.com>