Change backup directory for docker compatibility
Improve docker compatibility: Change the backup directory for ./nextcloud/data to reside within the host file system (basedir) so as to avoid moving the data folder back and forth between the docker container and the host file system. In situations where the nextcloud data takes up more than the free available space on the host system, this may lead to loosing some or all of the data.

Signed-off-by: MB-Finski <64466176+MB-Finski@users.noreply.github.com>