Update ncp-diag
I made the changes for checking if port 80 & 443 are open with the endpoint I've deployed. 
I changed the layout a little bit making it one row for IPv4 & one row for IPv6, this way it's a little bit more efficient since we only need to make 1 request to check both port 80 & 443 instead of 2, meaning the total requests goes down from 4 requests to 2 requests. 

Signed-off-by: Victor-ray, S. <12261439+ZendaiOwl@users.noreply.github.com>
Signed-off-by: Victor-ray, S <12261439+ZendaiOwl@users.noreply.github.com>