Improve performance of the "fix permissions" step in ncp-update-nc
The "fix permissions" step is really slow when the nextcloud folder contains a lot of data. In a docker setup the data directory is in nextcloud/data so the folder can be really big.
Even on my small test setup the "fix permissions" step took more than 20 minutes.
Using "chmod -R ..." reduces the time to  around 20 seconds.

Signed-off-by: Florian Wallner <asdf@walura.eu>